+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 * Fix non Texture2D textures that are the same reporting that they can't be combined.
 * Fix non Texture2D textures trying to combine with `null`.
 * Fix crash when vertex shader has no input struct.
+* Fix geometry shader input type to also allow for point & line.
 
 ## v1.4
 ### Features